code mới thêm

+++ src/main/java/Gwesty/Model/CreditCard.java
@@ -70,7 +70,7 @@ public class CreditCard {
         if (o == null || getClass() != o.getClass()) return false;
         CreditCard creditCard = (CreditCard) o;
         return Double.compare(balance, creditCard.balance) == 0 &&
-                Float.compare(cvv, creditCard.cvv) == 0 &&
+                Integer.compare(cvv, creditCard.cvv) == 0 &&
                 Objects.equals(number, creditCard.number) &&
                 Objects.equals(name, creditCard.name) &&
                 Integer.compare(month, creditCard.month) == 0 &&
